Update ci, in line with gitdb
This updates smmap's CI configuration in ways that are in line with
recent updates to gitdb's. In most cases there is no difference in
the changes, and the reason for the updates is more to avoid
confusing differences than from the value of the changes
themselves. In one case, there is a major difference.

The changes shall be listed below in the amended version of the
commit I open a PR for.